rename file and use `relativeToFile` function from `aurelia-path`
